+++ userController.php
@@ -4,8 +4,10 @@ include 'config/connectdb.php';
 
 $act = $_POST["act"];
 
-switch ($act) {
-    case 'add_user':
+if ($_SESSION['login'] == true) {
+
+    switch ($act) {
+        case 'add_user':
         $url = mysqli_real_escape_string($conn, $_POST["url"]);
         $nama = mysqli_real_escape_string($conn, $_POST["nama"]);
         $username = mysqli_real_escape_string($conn, $_POST["username"]);
@@ -16,46 +18,70 @@ switch ($act) {
         $tglupload = date('YmdHis');
         $file_name = basename($tglupload . $_FILES["file"]["name"]);
 
-        $target_dir = "images/fotoprofil/";
-        $target_file = $target_dir . $tglupload . $_FILES['file']['name'];
-        $imageFileType = strtolower(pathinfo($target_file, PATHINFO_EXTENSION));
-        $uploadOk = 1;
+        if ($_FILES["file"]["size"] != 0) {
+            $target_dir = "images/fotoprofil/";
+            $target_file = $target_dir . $tglupload . $_FILES['file']['name'];
+            $imageFileType = strtolower(pathinfo($target_file, PATHINFO_EXTENSION));
+            $uploadOk = 1;
 
-        //BATASI FILE CUMAN 5MB
-        if ($_FILES["file"]["size"] > 5485760) {
-            echo '<script type="text/javascript">alert("Your file too big, max 5mb"); </script>';
-            $uploadOk = 0;
-        }
-        if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif") {
-            echo '<script type="text/javascript">alert("Sorry, only JPG, JPEG, PNG & GIF files are allowed."); </script>';
-            $uploadOk = 0;
-        }
+                //BATASI FILE CUMAN 5MB
+            if ($_FILES["file"]["size"] > 5485760) {
+                echo '<script type="text/javascript">alert("Your file too big, max 5mb"); </script>';
+                $uploadOk = 0;
+            }
+            if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif") {
+                echo '<script type="text/javascript">alert("Sorry, only JPG, JPEG, PNG & GIF files are allowed."); </script>';
+                $uploadOk = 0;
+            }
 
-        if ($uploadOk == 0) {
-            echo '<script type="text/javascript">alert("Fail to insert new user"); </script>';
-            // if everything is ok, try to upload file
-        } else {
-            if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_file)) {
+            if ($uploadOk == 0) {
+                echo '<script type="text/javascript">alert("Fail to insert new user"); </script>';
+                    // if everything is ok, try to upload file
+            } else {
+                if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_file)) {
 
-                $sql = "INSERT INTO user (nama, username, password, role, kelas_id, ortu_id, foto)
-		VALUES ('$nama', '$username', '$password', '$role', $kelas, $ortu_id, '$file_name')";
-                if (mysqli_query($conn, $sql)) {
-                    echo '<script type="text/javascript">alert("Berhasil Tambah User Baru"); </script>';
-                    echo '<script type="text/javascript"> window.location = "' . $url . '" </script>';
-                } else {
-                    echo "Error: " . $sql . "<br>" . mysqli_error($conn);
+                    $sql = "INSERT INTO user (nama, username, password, role, kelas_id, ortu_id, foto)
+                    VALUES ('$nama', '$username', '$password', '$role', $kelas, $ortu_id, '$file_name')";
+                    if (mysqli_query($conn, $sql)) {
+                        echo '<script type="text/javascript">alert("Berhasil Tambah User Baru"); </script>';
+                        echo '<script type="text/javascript"> window.location = "' . $url . '" </script>';
+                    } else {
+                        echo "Error: " . $sql . "<br>" . mysqli_error($conn);
+                    }
                 }
+            }
+        } else {
+            $sql = "INSERT INTO user (nama, username, password, role, kelas_id, ortu_id)
+            VALUES ('$nama', '$username', '$password', '$role', $kelas, $ortu_id)";
+            if (mysqli_query($conn, $sql)) {
+                echo '<script type="text/javascript">alert("Berhasil Tambah User Baru"); </script>';
+                echo '<script type="text/javascript"> window.location = "' . $url . '" </script>';
+            } else {
+                echo "Error: " . $sql . "<br>" . mysqli_error($conn);
             }
         }
         break;
 
-    case 'edit_user':
+        case 'edit_user':
         $url = mysqli_real_escape_string($conn, $_POST["url"]);
         $user_id = mysqli_real_escape_string($conn, $_POST["user_id"]);
         $nama = mysqli_real_escape_string($conn, $_POST["nama"]);
         $password = mysqli_real_escape_string($conn, $_POST["password"]);
         $kelas = mysqli_real_escape_string($conn, $_POST["select-kelas"]);
         $ortu_id = mysqli_real_escape_string($conn, $_POST["select-ortu"]);
+
+        $q = "SELECT * FROM user WHERE id = '$user_id'";
+        $h = $conn->query($q);
+        if ($h->num_rows > 0) {
+            while ($r = $h->fetch_assoc()) {
+                $idkelas = $r['kelas_id'];
+                if ($kelas != $idkelas) {
+                    $q2 = "DELETE FROM relasi_user_matpel WHERE user_id = '$user_id'";
+                    mysqli_query($conn, $q2);
+                }
+            }
+        }
+
 
         if ($_FILES["file"]["size"] != 0) {
             $tglupload = date('YmdHis');
@@ -64,14 +90,14 @@ switch ($act) {
             $target_file = $target_dir . $tglupload . $_FILES['file']['name'];
             $uploadOk = 1;
 
-            //BATASI FILE CUMAN 5MB
+                //BATASI FILE CUMAN 5MB
             if ($_FILES["file"]["size"] > 5485760) {
                 echo "Sorry, your file is too large.";
                 $uploadOk = 0;
             }
             if ($uploadOk == 0) {
                 echo "Sorry, your file was not uploaded.";
-                // if everything is ok, try to upload file
+                    // if everything is ok, try to upload file
             } else {
 
                 if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_file)) {
@@ -80,7 +106,8 @@ switch ($act) {
                     if ($result->num_rows > 0) {
                         while ($row = $result->fetch_assoc()) {
                             $foto = $row['foto'];
-                            unlink('images/fotoprofil/' . $foto);
+                            if (isset($foto))
+                                unlink('images/fotoprofil/' . $foto);
                         }
                     }
                     $sql = "UPDATE user SET nama = '$nama', password = '$password', kelas_id = $kelas, ortu_id = $ortu_id, foto = '$file_name' WHERE id = '$user_id'";
@@ -98,11 +125,11 @@ switch ($act) {
             $sql = "UPDATE user SET nama = '$nama', password = '$password', kelas_id = $kelas, ortu_id = $ortu_id WHERE id = '$user_id'";
             if (mysqli_query($conn, $sql)) {
                 echo '<script type="text/javascript">alert("Berhasil edit user"); </script>';
-                // echo $nama; 
-                // echo $password;
-                // echo $kelas;
-                // echo $ortu_id;
-                // echo $user_id;
+                    // echo $nama; 
+                    // echo $password;
+                    // echo $kelas;
+                    // echo $ortu_id;
+                    // echo $user_id;
                 echo '<script type="text/javascript"> window.location = "' . $url . '" </script>';
             } else {
                 echo "Error: " . $sql . "<br>" . mysqli_error($conn);
@@ -110,7 +137,7 @@ switch ($act) {
         }
         break;
 
-    case 'delete_user':
+        case 'delete_user':
         $url = mysqli_real_escape_string($conn, $_POST["url"]);
         $dir = "images/fotoprofil";
         $user_id = mysqli_real_escape_string($conn, $_POST["user_id"]);
@@ -125,28 +152,32 @@ switch ($act) {
 
         $sql = "DELETE FROM user WHERE id = '$user_id'";
         if (mysqli_query($conn, $sql)) {
-            if (unlink($dir . '/' . $file)) {
-                echo '<script type="text/javascript">alert("Berhasil menghapus user"); </script>';
-                echo '<script type="text/javascript"> window.location = "' . $url . '" </script>';
+            if (isset($file)) {
+                if (unlink($dir . '/' . $file)) {
+                    echo '<script type="text/javascript">alert("Berhasil menghapus user"); </script>';
+                    echo '<script type="text/javascript"> window.location = "' . $url . '" </script>';
+                }
             }
         } else {
             echo "Error: " . $sql . "<br>" . mysqli_error($conn);
         }
         break;
 
-    case 'update_relasi':
+        case 'update_relasi':
         $url = mysqli_real_escape_string($conn, $_POST["url"]);
         $user_id = mysqli_real_escape_string($conn, $_POST["user_id"]);
 
         $sql = "DELETE FROM relasi_user_matpel WHERE user_id = '$user_id'";
         if (mysqli_query($conn, $sql)) {
-            foreach ($_POST['matpel'] as $matpel_id) {
-                $sql2 = "INSERT INTO relasi_user_matpel (user_id, matpel_id)
-		VALUES ('$user_id', '$matpel_id')";
-                if (mysqli_query($conn, $sql2)) {
-                    echo '<script type="text/javascript">alert("matpel id ' . $matpel_id . '"); </script>';
-                } else {
-                    echo "Error: " . $sql . "<br>" . mysqli_error($conn);
+            if (isset($_POST['matpel'])) {
+                foreach ($_POST['matpel'] as $matpel_id) {
+                    $sql2 = "INSERT INTO relasi_user_matpel (user_id, matpel_id)
+                    VALUES ('$user_id', '$matpel_id')";
+                    if (mysqli_query($conn, $sql2)) {
+                            // echo '<script type="text/javascript">alert("matpel id ' . $matpel_id . '"); </script>';
+                    } else {
+                        echo "Error: " . $sql . "<br>" . mysqli_error($conn);
+                    }
                 }
             }
             echo '<script type="text/javascript">alert("Berhasil edit relasi"); </script>';
@@ -155,5 +186,9 @@ switch ($act) {
             echo "Error: " . $sql . "<br>" . mysqli_error($conn);
         }
         break;
+    }
+} else {
+    echo '<script type="text/javascript">alert("Silahkan Login Terlebih Dahulu"); </script>';
+    echo '<script type="text/javascript"> window.location = "index.php" </script>';
 }
 ?>
